Factor out Point type to new sql.types
